Rename snap to ``octavia-diskimage-retrofit``
The snap is now strictly confined again, and we have added a few
more dependencies to it.

We are now at a point where we can iterate on changes to the
``libguestfs`` ``virt-dib`` command or indeed a new tool that
makes use of the libguestfs APIs if that turns out to make more
sense.