Build libguestfs fixed appliance and include it in snap
The upstream libguestfs tools expect to bootstrap a appliance from
system installed binaries.  This does not work easilly with a
strictly confined snap environment, so we distribute a pre-built
appliance instead.

This commit also fixes the compiled in path to the qemu executable.
The libguestfish configure scripts stores the full path to the
binary in path, and that happens to be the snapcraft staging area.

Temporarilly create a symlink in /snap in the build environment so
we can lure libguestfs configure script into testing and storing
a path that will be valid on a installed system.

Fixes up a couple of incorrect paths in qemus configflags.