Prune build- and stage- dependencies
Remove uneccessary build- and stage- packages from the ``qemu``
and ``libguestfs`` parts.

Initial iterations went out wide to get a PoC together quickly,
it's time to lean this thing up a bit.

Build ``supermin`` dependency ourselves instead of using
``libguestfs-tools`` as a build-package.  This is necessary to be
able to customize what goes into the fixed appliance at build
time.