Change location for the fixed appliance
As a consequence of building the fixed appliance from in-tree
tools instead of build-packages we also get the supermin tools
and the ``supermin.d`` directory in the snap.

The various libguestfs tools expects to use supermin to create a
appliance at runtime when ``supermin.d`` is present.

Add the fixed appliance at a different location to allow for using
the fixed appliance already built and included in the snap.