Fix ``QEMU`` build-environment variable
Commit a84d941dc7920c63bb1999c4b134a777a52ef76b changed how the
variable is built to accomodate multiple architecture support, but
failed to actually include the full path to the binary.

Not having the full path to the binary to use will make the
configure script pick up whatever it finds in the path at build
time, and this path is invalid on a installed system.