libguestfs: Change how path to QEMU binary is configured
Use ``--with-qemu`` to indicate where the installed binary will be.

Use ``QEMU`` build-environment variable to indicate where the
staged binary is for build-time configure tests.

Add patch for autgen.sh to make it honour ``NOCONFIGURE`` to avoid
running configure twice.

We still have to create /snap/NAME directory and symlink to the
stage directrory at build time to alleviate some assumptions made
in the build process when linking.