Resize image before retrofitting
Not all images have enough free space to add the necessary bits.
Increase the virtual size of the image with 1GB and run the
``growrootfs`` pre-install ``element``.

Passing the ``-r`` option will skip the resize operation.

Adds the ``growrootfs`` element for on-line resizing of root
filesystem at image build time.

Note that this addition differs from the upstream ``growroot``
element in that it does a build time on-line resize, whereas the
upstream ``growroot`` element adds init script for growing root
partition at first boot.  We'll let cloud-init defaults handle
that bit for us.