Increase the size of image to make room for updates
At this point in time attempting to retrofit a Focal image built
20200804 will fail due to the amount of updates being installed
exceeding the size of the root filesystem in the image.

We already resize the image and grow the root filesystem as part
of the process, increase the size we resize to to give more head
room.

Also drop the CI inspect target as it has been removed from
snapcraft. (ref snapcore/snapcraft/3c8cc28)

Closes-Bug: #1893697