Fix timesync issue for guestfs appliance on ARM
Before this change, when attempting to use the guestfs appliance
on ARM systems, many operations would fail due to the clock
showing some time around the Unix epoch (January 1st 1970).

This does not happen on other architectures as the timesync
between hypervisor and guest is built-in there.

Add the chrony package to the appliance and run it on startup to
fix this issue.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>